requirements.txt
astroid==2.2.5
certifi==2019.3.9
Django==2.2.4
ffmpeg-python==0.2.0
future==0.17.1
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysqlclient==1.4.4
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.1.0
pylint==2.3.1
PyMySQL==0.9.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
torch==1.1.0.post2
torchvision==0.3.0
tqdm==4.33.0
typed-ast==1.4.0
wrapt==1.11.2
youtube-dl==2019.8.13